parsers/teachers_timetable/requirements.txt
selenium==4.24.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
lxml==5.3.0
requests==2.31.0